requirements.txt
numpy==2.2.6
scikit-learn==1.7.0
pandas==2.3.1
matplotlib==3.10.3
scipy==1.15.3
opencv-python==4.12.0.88
transformers==4.53.2
jieba==0.42.1
pillow==11.3.0
beautifulsoup4==4.13.4
xgboost-cpu==3.0.2
spacy==3.8.7
plotly==6.2.0
onnx==1.18.0
onnxruntime==1.22.1
Flask==3.1.1